% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jotify-package.R
\docType{package}
\name{jotify-package}
\alias{jotify}
\alias{jotify-package}
\title{jotify: Charts and Utilities for the Moodring Web Application}
\description{
This package holds boutique chart and utilities functions used throughout the Moodring web application, an app that provides real-time analysis of J.J.'s recent Spotify listening history.
}
\author{
\strong{Maintainer}: J.J. Moncus \email{jjmoncus706@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
