% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_jotify}
\alias{theme_jotify}
\title{ggplot theme for jotify graphics}
\usage{
theme_jotify(
  head_size = 12,
  body_size = 9,
  family = "Franklin",
  x_axis = TRUE,
  y_axis = TRUE,
  legend = TRUE
)
}
\description{
ggplot theme for jotify graphics
}
